Remove vtables from JSON library
