+++ BUILD.gn
@@ -690,12 +690,8 @@ config("grpc_config") {
         "src/core/lib/iomgr/wakeup_fd_posix.h",
         "src/core/lib/json/json.cc",
         "src/core/lib/json/json.h",
-        "src/core/lib/json/json_common.h",
         "src/core/lib/json/json_reader.cc",
-        "src/core/lib/json/json_reader.h",
-        "src/core/lib/json/json_string.cc",
         "src/core/lib/json/json_writer.cc",
-        "src/core/lib/json/json_writer.h",
         "src/core/lib/security/context/security_context.cc",
         "src/core/lib/security/context/security_context.h",
         "src/core/lib/security/credentials/alts/alts_credentials.cc",
@@ -1528,12 +1524,8 @@ config("grpc_config") {
         "src/core/lib/iomgr/wakeup_fd_posix.h",
         "src/core/lib/json/json.cc",
         "src/core/lib/json/json.h",
-        "src/core/lib/json/json_common.h",
         "src/core/lib/json/json_reader.cc",
-        "src/core/lib/json/json_reader.h",
-        "src/core/lib/json/json_string.cc",
         "src/core/lib/json/json_writer.cc",
-        "src/core/lib/json/json_writer.h",
         "src/core/lib/profiling/timers.h",
         "src/core/lib/slice/b64.cc",
         "src/core/lib/slice/b64.h",
