+++ gRPC-Core.podspec
@@ -739,12 +739,8 @@ Pod::Spec.new do |s|
                       'src/core/lib/iomgr/wakeup_fd_posix.h',
                       'src/core/lib/json/json.cc',
                       'src/core/lib/json/json.h',
-                      'src/core/lib/json/json_common.h',
                       'src/core/lib/json/json_reader.cc',
-                      'src/core/lib/json/json_reader.h',
-                      'src/core/lib/json/json_string.cc',
                       'src/core/lib/json/json_writer.cc',
-                      'src/core/lib/json/json_writer.h',
                       'src/core/lib/profiling/basic_timers.cc',
                       'src/core/lib/profiling/stap_timers.cc',
                       'src/core/lib/profiling/timers.h',
@@ -1224,9 +1220,6 @@ Pod::Spec.new do |s|
                               'src/core/lib/iomgr/wakeup_fd_pipe.h',
                               'src/core/lib/iomgr/wakeup_fd_posix.h',
                               'src/core/lib/json/json.h',
-                              'src/core/lib/json/json_common.h',
-                              'src/core/lib/json/json_reader.h',
-                              'src/core/lib/json/json_writer.h',
                               'src/core/lib/profiling/timers.h',
                               'src/core/lib/security/context/security_context.h',
                               'src/core/lib/security/credentials/alts/alts_credentials.h',
