+++ grpc.gemspec
@@ -651,12 +651,8 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/iomgr/wakeup_fd_posix.h )
   s.files += %w( src/core/lib/json/json.cc )
   s.files += %w( src/core/lib/json/json.h )
-  s.files += %w( src/core/lib/json/json_common.h )
   s.files += %w( src/core/lib/json/json_reader.cc )
-  s.files += %w( src/core/lib/json/json_reader.h )
-  s.files += %w( src/core/lib/json/json_string.cc )
   s.files += %w( src/core/lib/json/json_writer.cc )
-  s.files += %w( src/core/lib/json/json_writer.h )
   s.files += %w( src/core/lib/profiling/basic_timers.cc )
   s.files += %w( src/core/lib/profiling/stap_timers.cc )
   s.files += %w( src/core/lib/profiling/timers.h )
