+++ grpc.gyp
@@ -347,7 +347,6 @@
         'src/core/lib/iomgr/wakeup_fd_posix.cc',
         'src/core/lib/json/json.cc',
         'src/core/lib/json/json_reader.cc',
-        'src/core/lib/json/json_string.cc',
         'src/core/lib/json/json_writer.cc',
         'src/core/lib/slice/b64.cc',
         'src/core/lib/slice/percent_encoding.cc',
@@ -759,7 +758,6 @@
         'src/core/lib/iomgr/wakeup_fd_posix.cc',
         'src/core/lib/json/json.cc',
         'src/core/lib/json/json_reader.cc',
-        'src/core/lib/json/json_string.cc',
         'src/core/lib/json/json_writer.cc',
         'src/core/lib/slice/b64.cc',
         'src/core/lib/slice/percent_encoding.cc',
@@ -1022,7 +1020,6 @@
         'src/core/lib/iomgr/wakeup_fd_posix.cc',
         'src/core/lib/json/json.cc',
         'src/core/lib/json/json_reader.cc',
-        'src/core/lib/json/json_string.cc',
         'src/core/lib/json/json_writer.cc',
         'src/core/lib/slice/b64.cc',
         'src/core/lib/slice/percent_encoding.cc',
@@ -1261,7 +1258,6 @@
         'src/core/lib/iomgr/wakeup_fd_posix.cc',
         'src/core/lib/json/json.cc',
         'src/core/lib/json/json_reader.cc',
-        'src/core/lib/json/json_string.cc',
         'src/core/lib/json/json_writer.cc',
         'src/core/lib/slice/b64.cc',
         'src/core/lib/slice/percent_encoding.cc',
@@ -1690,7 +1686,6 @@
         'src/core/lib/iomgr/wakeup_fd_posix.cc',
         'src/core/lib/json/json.cc',
         'src/core/lib/json/json_reader.cc',
-        'src/core/lib/json/json_string.cc',
         'src/core/lib/json/json_writer.cc',
         'src/core/lib/slice/b64.cc',
         'src/core/lib/slice/percent_encoding.cc',
@@ -2044,7 +2039,6 @@
         'src/core/lib/iomgr/wakeup_fd_posix.cc',
         'src/core/lib/json/json.cc',
         'src/core/lib/json/json_reader.cc',
-        'src/core/lib/json/json_string.cc',
         'src/core/lib/json/json_writer.cc',
         'src/core/lib/slice/b64.cc',
         'src/core/lib/slice/percent_encoding.cc',
