+++ src/python/grpcio/grpc_core_dependencies.py
@@ -314,7 +314,6 @@ CORE_SOURCE_FILES = [
     'src/core/lib/iomgr/wakeup_fd_posix.cc',
     'src/core/lib/json/json.cc',
     'src/core/lib/json/json_reader.cc',
-    'src/core/lib/json/json_string.cc',
     'src/core/lib/json/json_writer.cc',
     'src/core/lib/profiling/basic_timers.cc',
     'src/core/lib/profiling/stap_timers.cc',
